Introduced SFT_UChar type for UTF-32 codepoints
We needed an integer type for Unicode codepoints that is not as platform-dependent as 'unsigned long'.
Theoretically we could have used 'char32_t' from 'uchar.h'.
But many platforms, event those with official C11 support, don't feature this header.
So instead, we define our own typedef in such a way that is guaranteed to be the same type as char32_t
on platforms that support it.